requirements.txt
# Finora — متطلبات التشغيل
Flask
Flask-SQLAlchemy
python-dotenv
werkzeug
reportlab
pandas
gunicorn
Flask-Compress
